requirements.txt
python-dotenv>=1.0,<2
pydantic>=2.0,<3
openai>=2.0,<3
modal>=1.0,<2
fastapi>=0.100,<1
feedparser>=6.0,<7
